calcul montant && montant retard

+++ src/main/java/com/fstg/taxesejour/infrastructure/entity/TaxeSejourTrim.java
@@ -20,20 +20,13 @@ public class TaxeSejourTrim extends AbstractEntity {
     int numTrim;
     int annee;
     String ref;
-    BigDecimal nombreNuit;
+    Long nombreNuit;
     BigDecimal montant;
     BigDecimal montantRetard;
     @ManyToOne()
     TaxeSejourAnnuele taxeSejourAnnuele;
     @Temporal(TemporalType.DATE)
     Date datePresentation;
-//
-//    @Temporal(TemporalType.DATE)
-//    @Transient
-//    Date expectedDatePresentation;
-
-    //    @Temporal(TemporalType.DATE)
-//    Date dateValidation;
     @ManyToOne()
     TauxTaxeSejour tauxTaxeSejour;
     @OneToOne(optional = true)
